refactoring for a cleaner cli... I think
